Moving files metadata parsers logic to parsers.py
